src/main/java/hu/wirthandras/owl/domain/Translation.java
package hu.wirthandras.owl.domain;

public class Translation implements Comparable<Translation> {
	
	private int id;
	private String base;
	private String translation;
	private int occurences;
	private int answered;
	private int correct;
	private int percent;
	
	public Translation() {
	
	}

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getBase() {
		return base;
	}

	public void setBase(String base) {
		this.base = base;
	}

	public String getTranslation() {
		return translation;
	}

	public void setTranslation(String translation) {
		this.translation = translation;
	}

	public int getOccurences() {
		return occurences;
	}

	public void setOccurences(int occurences) {
		this.occurences = occurences;
	}

	public int getAnswered() {
		return answered;
	}

	public void setAnswered(int answered) {
		this.answered = answered;
	}

	public int getCorrect() {
		return correct;
	}

	public void setCorrect(int correct) {
		this.correct = correct;
	}

	@Override
	public int compareTo(Translation o) {
		return Integer.compare(percent, o.getPercent());
	}

	public int getPercent() {
		return percent;
	}

	public void setPercent(int percent) {
		this.percent = percent;
	}	
	
}
